Apply atomic design completely
