+++ src/pages/boards/index.tsx
@@ -1,18 +1,9 @@
 import * as constants from '@/src/constants';
 
+import BoardsIndexView from '@/src/components/templates/boards/index/IndexView';
 import { GetServerSideProps } from 'next';
-import Grid from '@/src/components/atoms/grid/Grid';
-import Image from '@/src/components/atoms/image/Image';
 import type { NextPage } from 'next';
-import PageOverview from '@/src/components/molecules/pageOverview/PageOverview';
-import Section from '@/src/components/molecules/section/Section';
-import Thumbnail from '@/src/components/molecules/thumbnail/Thumbnail';
-import WidthLimiter from '@/src/components/atoms/widthLimiter/WidthLimiter';
-import WithCommonPreferences from '@/src/components/WithCommonPreferences';
 import { getBoardsByNames } from '@/src/backend/dao/board';
-import styled from '@emotion/styled';
-import { useRouter } from 'next/router';
-import { useTranslation } from 'next-i18next';
 import { withAuthServerSideProps } from '@/src/utils/server-side';
 
 export const getServerSideProps: GetServerSideProps = withAuthServerSideProps('user',
@@ -50,17 +41,6 @@ export const getServerSideProps: GetServerSideProps = withAuthServerSideProps('u
   }
 );
 
-const StyledPageOverviewImageWrap = styled(Image)`
-  ${(props: any) => `
-    display: block;
-    width: 100%;
-    max-height: 480px;
-    max-width: 480px;
-    margin: 0 auto;
-    vertical-align: top;
-  `}
-`;
-
 interface Board {
   id: number;
   name: string;
@@ -75,52 +55,10 @@ interface PageProps {
   boards: Board[]
 }
 
-const BoardsIndex: NextPage<PageProps> = ({
-  callbackUrl,
-  currentUser,
-  boards
-}) => {
-  const router = useRouter();
-  const { t } = useTranslation('common');
+const BoardsIndex: NextPage<PageProps> = props => {
   return (
-    <WidthLimiter>
-      <Section variant='transparent'>
-        <PageOverview
-          firstChildren={<>
-            <StyledPageOverviewImageWrap
-              src='https://image.freepik.com/free-vector/tiny-characters-sitting-laptop-with-lorem-ipsum-title_74855-20389.jpg?w=996'
-            />
-          </>}
-          secondChildren={<>
-            Hello
-          </>}
-        />
-        <Grid
-          column={{
-            default: 1,
-            gtePhablet: 2,
-            gteTablet: 3,
-          }}
-        >
-          {
-            boards
-            .sort((lhs, rhs) => t(lhs.name) > t(rhs.name) ? 1 : -1)
-            .map((board: Board, index: number) => {
-              return <Thumbnail
-                key={index}
-                imageUrl={board.image.address}
-                title={t(board.name)}
-                onClick={() => router.push({
-                  pathname: '/boards/[id]',
-                  query: { id: board.id },
-                })}
-              />
-            })
-          }
-        </Grid>
-      </Section>
-    </WidthLimiter>
+    <BoardsIndexView {...props} />
   );
 }
 
-export default WithCommonPreferences(BoardsIndex);
+export default BoardsIndex;
