+++ src/pages/home/index.tsx
@@ -1,10 +1,7 @@
-import Card from '@/src/components/atoms/card/Card';
 import { GetServerSideProps } from 'next';
+import HomeIndexView from '@/src/components/templates/home/index/IndexView';
 import type { NextPage } from 'next';
-import Section from '@/src/components/molecules/section/Section';
-import WidthLimiter from '@/src/components/atoms/widthLimiter/WidthLimiter';
 import WithCommonPreferences from '@/src/components/WithCommonPreferences';
-import { useTranslation } from 'next-i18next';
 import { withAuthServerSideProps } from '@/src/utils/server-side';
 
 export const getServerSideProps: GetServerSideProps = withAuthServerSideProps('all');
@@ -14,14 +11,9 @@ interface PageProps {
   callbackUrl: string,
 }
 
-const HomeIndex: NextPage<PageProps> = ({ currentUser, callbackUrl }) => {
-  const { t } = useTranslation('common');
+const HomeIndex: NextPage<PageProps> = props => {
   return (
-    <WidthLimiter>
-      <Section>
-        {t('_helloWorld')}
-      </Section>
-    </WidthLimiter>
+    <HomeIndexView {...props} />
   );
 }
