+++ src/pages/sandbox/index.tsx
@@ -1,14 +1,7 @@
-import Button from '@/src/components/atoms/button/Button';
-import Dialog from '@/src/components/molecules/dialog/Dialog';
 import { GetServerSideProps } from 'next';
-import Loader from '@/src/components/atoms/loader/Loader';
 import type { NextPage } from 'next';
-import Section from '@/src/components/molecules/section/Section';
-import WidthLimiter from '@/src/components/atoms/widthLimiter/WidthLimiter';
+import SandboxIndexView from '@/src/components/templates/sandbox/index/IndexView';
 import WithCommonPreferences from '@/src/components/WithCommonPreferences';
-import { useRouter } from 'next/router';
-import { useState } from 'react';
-import { useTranslation } from 'next-i18next';
 import { withAuthServerSideProps } from '@/src/utils/server-side';
 
 export const getServerSideProps: GetServerSideProps = withAuthServerSideProps('all');
@@ -21,49 +14,9 @@ interface PageProps {
   setIsDark: (isDark: boolean) => {};
 }
 
-const SandboxIndex: NextPage<PageProps> = ({
-  callbackUrl,
-  currentUser,
-  isDark,
-  setIsDark,
-}) => {
-  const router = useRouter();
-  const { t } = useTranslation('common');
-  const [open, setOpen] = useState(false);
+const SandboxIndex: NextPage<PageProps> = props => {
   return (
-    <WidthLimiter>
-      <Section>
-        {t('_helloWorld')}
-        <Button
-          variant='filled'
-          palette='primary'
-          onClick={() => router.push('/sandbox', '/sandbox', { locale: router.locale === 'en' ? 'ko' : 'en' })}
-        >
-          {`change-locale`}
-        </Button>
-        <Button
-          variant='filled'
-          palette='primary'
-          onClick={() => setIsDark(!isDark)}
-        >
-          {isDark ? 'Dark' : 'Light'}
-        </Button>
-        <Button
-          onClick={() => setOpen(true)}
-        >
-          Toggle Modal
-        </Button>
-        <Loader />
-        <Dialog
-          open={open}
-          onClickBackdrop={() => setOpen(false)}
-          title={'테스트 제목입니다.'}
-          content={'테스트 본문입니다.'}
-          onYes={() => {}}
-          onNo={() => {}}
-        />
-      </Section>
-    </WidthLimiter>
+    <SandboxIndexView {...props} />
   );
 }
